chore(stop_filter): move include files
