+++ localization/autoware_stop_filter/test/test_stop_filter.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware/stop_filter/stop_filter.hpp"
+#include "../src/stop_filter.hpp"
 
 #include <gtest/gtest.h>
